tests/wpt/web-platform-tests/workers/semantics/encodings/002.html

<!doctype html>
<title>encoding, shared worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new SharedWorker('002.js', '');
  worker.port.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '\u00e5');
    this.done();
  });
});
</script>
